requirements.txt

groq
python-dotenv